boxworks: grab bag of changes
- adds traits for the line breaker and font width provider so that alternative
  implementations can be made.

- adds texlang support for all parameters in the line breaker

- reorganizes the boxworks-texlang crate to give each boxworks crate its own
  module.

- implements \sfcode.

+++ crates/boxworks-knuthplass/src/debug.rs
@@ -4,7 +4,7 @@ use boxworks::ds;
 
 pub struct FeasibleBreakpoint {
     pub elem_index: usize,
-    pub badness: i64,
+    pub badness: i32,
     pub penalty: i32,
     pub demerits: i32,
     pub previous_node_index: usize,
